Refactored BTLE server;
