+++ app/src/main/java/com/dataart/btle_android/btle_gateway/server/ConnectionOperation.java
@@ -1,0 +1,9 @@
+package com.dataart.btle_android.btle_gateway.server;
+
+import com.dataart.btle_android.btle_gateway.gatt_callbacks.DeviceConnection;
+
+interface ConnectionOperation {
+    void call(DeviceConnection connection);
+
+    void fail(String message);
+}
