+++ app/src/main/java/com/dataart/btle_android/btle_gateway/server/DeviceOperation.java
@@ -1,0 +1,9 @@
+package com.dataart.btle_android.btle_gateway.server;
+
+import android.bluetooth.BluetoothDevice;
+
+interface DeviceOperation {
+    void call(BluetoothDevice device);
+
+    void fail(String message);
+}
