+++ app/src/main/java/com/dataart/btle_android/btle_gateway/server/DiscoveredDeviceListener.java
@@ -1,0 +1,7 @@
+package com.dataart.btle_android.btle_gateway.server;
+
+import android.bluetooth.BluetoothDevice;
+
+interface DiscoveredDeviceListener {
+    void onDiscoveredDevice(BluetoothDevice device);
+}
